+++ naga/src/back/hlsl/mod.rs
@@ -793,7 +793,8 @@ pub fn supported_capabilities() -> crate::valid::Capabilities {
         | Caps::ACCELERATION_STRUCTURE_BINDING_ARRAY
         // No CLIP_DISTANCES
         // No CULL_DISTANCE
-        | Caps::STORAGE_TEXTURE_16BIT_NORM_FORMATS
+        | Caps::TEXTURE_FORMATS_TIER1
+        | Caps::TEXTURE_FORMATS_TIER2
         | Caps::MULTIVIEW
         // No EARLY_DEPTH_TEST
         | Caps::MULTISAMPLED_SHADING